requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dateparser==1.1.8
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
flaskwebgui==1.0.4
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
multidict==6.0.4
packaging==23.1
psutil==5.9.4
pycryptodome==3.17
python-binance==1.0.17
python-dateutil==2.8.2
python-engineio==4.4.0
python-socketio==5.8.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
websocket-client==1.5.1
websockets==11.0.1
Werkzeug==2.2.3
WTForms==3.0.1
yarl==1.8.2
